ai-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
Pillow==10.1.0
numpy==1.24.3
gunicorn==21.2.0
python-multipart==0.0.6
Werkzeug==2.3.7